refac(flaginfo): put parsing sections in its own fils
